fix #1

+++ OnlineShop/Services/BrandService.cs
@@ -12,7 +12,7 @@ public interface IBrandService
     Task<IEnumerable<BrandDto>> GetAllBrandsAsync();
     Task<BrandDto> GetBrandByIdAsync(int brandId);
     Task<BrandDto> CreateBrandAsync(BrandAdd dto);
-    Task<BrandDto> UpdateBrandAsync(int brandId, BrandUpdate dto);
+    Task<BrandDto> UpdateBrandNameAsync(int brandId, BrandUpdate dto);
     Task<BrandDto> DeleteBrandAsync(int brandId);
 }
 
@@ -32,34 +32,37 @@ public class BrandService : IBrandService
 
     public async Task<BrandDto> GetBrandByIdAsync(int brandId)
     {
-        //var brand = await _context.Brands.ProjectToType<BrandDto>()
-        //    .FirstOrDefaultAsync(b => b.Id == brandId);
-
-        var brand = await _context.Brands.Include(b => b.Products).FirstOrDefaultAsync(b => b.Id == brandId);
-
+        var brand = await _context.Brands.ProjectToType<BrandDto>()
+            .FirstOrDefaultAsync(b => b.Id == brandId);
 
         if (brand == null)
             throw new NotFoundException(ExceptionMessages.BrandNotFound);
 
-        return brand.AdaptToDto();
+        return brand;
     }
 
     public async Task<BrandDto> CreateBrandAsync(BrandAdd dto)
     {
+        var candidate = await _context.Brands.FirstOrDefaultAsync(b => b.Name == dto.Name);
+        if (candidate is not null)
+            throw new AlreadyExistsException(ExceptionMessages.BrandAlreadyExists);
+
         var brand = dto.AdaptToBrand();
         await _context.Brands.AddAsync(brand);
         await _context.SaveChangesAsync();
-        var existingBrand = await _context.Brands.ProjectToType<BrandDto>()
-            .FirstAsync(b => b.Id == brand.Id);
-        return existingBrand;
+        return brand.AdaptToDto();
     }
 
-    public async Task<BrandDto> UpdateBrandAsync(int brandId, BrandUpdate dto)
+    public async Task<BrandDto> UpdateBrandNameAsync(int brandId, BrandUpdate dto)
     {
         var existingBrand = await _context.Brands.FindAsync(brandId);
 
-        if (existingBrand == null)
+        if (existingBrand is null)
             throw new NotFoundException(ExceptionMessages.BrandNotFound);
+
+        var candidate = await _context.Brands.FirstOrDefaultAsync(b => b.Name == dto.Name);
+        if (candidate is not null)
+            throw new AlreadyExistsException(ExceptionMessages.BrandAlreadyExists);
 
         existingBrand.Name = dto.Name;
         
@@ -69,10 +72,14 @@ public class BrandService : IBrandService
 
     public async Task<BrandDto> DeleteBrandAsync(int brandId)
     {
-        var existingBrand = await _context.Brands.FindAsync(brandId);
+        var existingBrand = await _context.Brands.Include(b => b.Products)
+            .FirstOrDefaultAsync(b => b.Id == brandId);
 
         if (existingBrand == null)
             throw new NotFoundException(ExceptionMessages.BrandNotFound);
+
+        if (existingBrand.Products.Count > 0)
+            throw new InvalidOperationException(ExceptionMessages.BrandHasProducts);
 
         _context.Brands.Remove(existingBrand);
         await _context.SaveChangesAsync();
